Experience is a double instead of integer as required by vanilla implementations.

+++ src/main/java/org/spongepowered/api/entity/ExperienceOrb.java
@@ -34,13 +34,13 @@ public interface ExperienceOrb extends Entity {
      *
      * @return Amount of experience
      */
-    public int getExperience();
+    public double getExperience();
 
     /**
      * Sets how much experience will be added to the player on pickup.
      *
      * @param experience The new amount of experience
      */
-    public void setExperience(int experience);
+    public void setExperience(double experience);
 
 }
